Handle argument register cyclic dependencies in x86 call gen
e.g.
	$g = i4()
	$use = void(...)
	$f = void()
	{
		$1 = alloca i4
		$2 = alloca i8
		$3 = alloca [i8 x 2]
		$4 = alloca i4
		$5 = alloca i8

		$a = call $g()
		$b = call $g()
		$c = call $g()
		$d = call $g()

		call $use($a, $b, $c, $d)

		ret void
	}

Here $a, $b, $c and $d could end up in a cycle where each is already
in an argument register, so there are none free to move into. In
this case we now use the scratch-register (guaranteed not to be an
argument register) to hold $a, move the rest, then move $a's value
from the scratch into the correct argument register.